remove old version of files
there was a flaw in the merge due to mismatched filenames. however, I can simply replace the index page php and js that were on this branch with the new versions, so there is nothing we needed from the index php and js from this tableview branch. All newer commits that were important to keep were in tableview.php