add combobox duplication workaround attempt
calling the combobox method after the combobox has already been initialized causes weird bugs, e.g. elements get duplicated (probably because it tries to act on the select tag that is still present). This issue occurs when we want to reload the msa options. However, if we don't call the combobox method after updating the options in the select tag, then the options don't correctly update/display in the combobox. The workaround that I implemented is this: at the beginning of loadMSA() we clear the non-static contents of the viewMsaGeo div (where I've put the msa dropdown/combobox) and replace it with a set starting version that is pre-conversion to combobox.