fix bug: msaList non-responsive to suppression
bug behavior: msaList was not triggering (what is now) the respondToFirstDD() on change. (respondToFirstDD() is what used to be an anonymous function that was triggered on change for $("#firstLevelGeoList, #firstLevelGeoListAlt, #firstLevelGeoListAlt2, #firstLevelGeoListAlt3"), but I had to name it in order to rebind it to msaList separate from the aforementioned geo lists.)
solution: on-change function just needed to be bound again to msaList after msaList gets recreated. (msaList needs to be recreated on every loadMsa() call, so that combobox can be called on it again, in order to update the combobox's options without duplicating combobox elements)

I also replaced the repetitive instances of the text "msaList"/"#msaList" in the code with a variable, ensuring that the correct element is referenced throughout the function to avoid mismatches in case in the future there are multiple msa lists